+++ src/styles/themes/color.ts
@@ -1,0 +1,26 @@
+export const COLOR = {
+  common: {
+    white: '#FFF',
+    black: '#000',
+    gray: {
+      900: '#202020',
+      800: '#3B3B3B',
+      700: '#5C5C5C',
+      600: '#797979',
+      500: '#A0A0A0',
+      400: '#C9C9C9',
+      300: '#E5E5E5',
+      200: '#EFEFEF',
+      100: '#F7F7F7',
+      50: '#FBFBFB',
+    },
+  },
+  success: '#1DCB63',
+  error: '#CB1D3C',
+  alert: {
+    100: '#DA1E28',
+    70: '#99151C',
+    40: '#660000',
+    20: '#2C0608',
+  },
+} as const
