+++ src/styles/themes/font.size.ts
@@ -1,0 +1,8 @@
+export const FONT_SIZE = {
+  tiny: '1.4rem',
+  small: '1.8rem',
+  medium: '2.0rem',
+  large: '2.4rem',
+  big: '3.2rem',
+  huge: '4.8rem',
+} as const
