+++ src/styles/themes/media.ts
@@ -1,0 +1,5 @@
+export const MEDIA = {
+  mobile: '414px',
+  tablet: '820px',
+  laptop: '1440px',
+} as const
