+++ src/styles/themes/palette.ts
@@ -1,0 +1,26 @@
+export const PALETTE = {
+  yellow: {
+    100: '#FFCA5A',
+    70: '#CCA148',
+    40: '#997936',
+    20: '#665124',
+  },
+  blue: {
+    100: '#1E8CFA',
+    70: '#1562AF',
+    40: '#0C3864',
+    20: '#061C32',
+  },
+  purple: {
+    100: '#54007B',
+    70: '#7A00B3',
+    40: '#460066',
+    20: '#230033',
+  },
+  orange: {
+    100: '#FF7A00',
+    70: '#B35500',
+    40: '#663100',
+    20: '#331800',
+  },
+} as const
